막대그래프 컴포넌트 분리, 마진적용
